Rename useSortState and usePaginationState to useLegacy*State
